Refactoring: move helpers to separate directory
