+++ helpers/parseAndFindVarsInProgram.js
