+++ lib/findVariablesInTemplate.js
@@ -9,7 +9,7 @@ const findVariablesInTemplate = (template) => {
 
     tokens.forEach(token => visitors(state, token, template))
 
-    return state.variables
+    return state.getVariables()
   }
 
   return []
