+++ lib/state.js
@@ -1,4 +1,4 @@
-const pushUniqueVariable = require('./pushUniqueVariable')
+const pushUniqueVariable = require('../helpers/pushUniqueVariable')
 
 function getVariablesInContext(context, column) {
   return Object.keys(context)
@@ -9,8 +9,15 @@ function getVariablesInContext(context, column) {
 
 module.exports = class State {
   constructor() {
+    // Array of variable Nodes
     this.variables = []
+
+    // Map of variable names
     this.context = {}
+  }
+
+  getVariables() {
+    return this.variables
   }
 
   addVariables(column, variables) {
