+++ lib/visitors.js
@@ -1,5 +1,5 @@
-const parseAndFindVarsInProgram = require('./parseAndFindVarsInProgram')
-const buildVariable = require('./buildVariable')
+const parseAndFindVarsInProgram = require('../helpers/parseAndFindVarsInProgram')
+const { buildVariable, findVariable } = require('../helpers/variables')
 
 function isComponent(string) {
   return /^[A-Z].*$/.test(string)
@@ -11,7 +11,7 @@ function isSpreadOperator(string) {
 
 module.exports = function visitors(state, token, template) {
   const location = token.loc
-  const locs = template.split('\n')
+  const source = template.split('\n').slice(location.start.line - 1, location.end.line)
 
   switch (token.type) {
     case 'outdent': {
@@ -22,14 +22,7 @@ module.exports = function visitors(state, token, template) {
 
     case 'tag': {
       if (isComponent(token.val)) {
-        const source = locs.slice(location.start.line - 1, location.end.line)
-        const name = token.val
-
-        const line = location.start.line
-        const valueStartsAt = location.start.column
-        const columnStart = valueStartsAt + source[0].substring(valueStartsAt).indexOf(name)
-        const columnEnd = columnStart + name.length
-        const variable = buildVariable(token.val, [line, columnStart], [line, columnEnd])
+        const variable = findVariable(token.val, location, source)
 
         state.addVariables(location.start.column, [variable])
       }
@@ -45,8 +38,6 @@ module.exports = function visitors(state, token, template) {
         // Pug can return a boolean variable for attribute value, but babel
         // can parses only strings
         : String(token.val)
-
-      const source = locs.slice(location.start.line - 1, location.end.line)
 
       const usedVariables = parseAndFindVarsInProgram(normalizedValue)
         .used
@@ -74,17 +65,9 @@ module.exports = function visitors(state, token, template) {
 
     case 'interpolated-code':
     case 'if': {
-      const source = locs.slice(location.start.line - 1, location.end.line)
       const usedVariables = parseAndFindVarsInProgram(token.val)
         .used
-        .map((name) => {
-          const line = location.start.line
-          const valueStartsAt = location.start.column
-          const columnStart = valueStartsAt + source[0].substring(valueStartsAt).indexOf(name)
-          const columnEnd = columnStart + name.length
-
-          return buildVariable(name, [line, columnStart], [line, columnEnd])
-        })
+        .map(name => findVariable(name, location, source))
 
       state.addVariables(location.start.column, usedVariables)
 
@@ -96,17 +79,9 @@ module.exports = function visitors(state, token, template) {
 
       state.addToContext(location.start.column, variables.defined)
 
-      const source = locs.slice(location.start.line - 1, location.end.line)
       const usedVariables = variables
         .used
-        .map((name) => {
-          const line = location.start.line
-          const valueStartsAt = location.start.column
-          const columnStart = valueStartsAt + source[0].substring(valueStartsAt).indexOf(name)
-          const columnEnd = columnStart + name.length
-
-          return buildVariable(name, [line, columnStart], [line, columnEnd])
-        })
+        .map(name => findVariable(name, location, source))
 
       state.addVariables(location.start.column, usedVariables)
 
@@ -115,23 +90,14 @@ module.exports = function visitors(state, token, template) {
 
     case 'each': {
       const defined = [token.val, token.key]
-      const variables = parseAndFindVarsInProgram(token.code)
 
       // We need to increase column because 'each' creates own scope and
       // defines variables inside it
       state.addToContext(location.start.column + 1, defined)
 
-      const source = locs.slice(location.start.line - 1, location.end.line)
-      const usedVariables = variables
+      const usedVariables = parseAndFindVarsInProgram(token.code)
         .used
-        .map((name) => {
-          const line = location.start.line
-          const valueStartsAt = location.start.column
-          const columnStart = valueStartsAt + source[0].substring(valueStartsAt).indexOf(name)
-          const columnEnd = columnStart + name.length
-
-          return buildVariable(name, [line, columnStart], [line, columnEnd])
-        })
+        .map(name => findVariable(name, location, source))
 
       state.addVariables(location.start.column, usedVariables)
